requirements.txt
tensorflow-cpu==2.8.0
Flask==2.0.2
gunicorn==20.1.0
sklearn==0.0
pandas==1.3.4
